requirements.txt
flask
ipykernel
python-dotenv
groq
streamlit
watchdog
werkzeug